Fixed bullet "body already added to space" bug!
(I think!)

I added some debug logging and discovered that this
was caused by close() being called on the bullet twice.
close() would (eventually) add the bullet to the freelist,
and since the freelist is a list we'd eventually pull both
instances off the freelist at the same time, which meant
adding to he PyMunk space twice, boom.

Maybe PyMunk is the culprit?  Colliding the bullet twice
in a single frame or something?  Anyway, setting a flag
so it ignores redundant closes seems to have fixed it.